requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
asttokens=2.0.5=pyhd8ed1ab_0
backcall=0.2.0=pyh9f0ad1d_0
backports=1.0=py_2
backports.functools_lru_cache=1.6.4=pyhd8ed1ab_0
ca-certificates=2021.10.8=h5b45459_0
certifi=2021.10.8=py39hcbf5309_2
charset-normalizer=2.0.12=pypi_0
colorama=0.4.4=pyhd3eb1b0_0
cudatoolkit=11.3.1=h59b6b97_2
cycler=0.11.0=pypi_0
debugpy=1.6.0=py39h415ef7b_0
decorator=5.1.1=pyhd8ed1ab_0
entrypoints=0.4=pyhd8ed1ab_0
executing=0.8.3=pyhd8ed1ab_0
fonttools=4.33.3=pypi_0
idna=3.3=pypi_0
ipykernel=6.13.0=py39h832f523_0
ipython=8.3.0=py39hcbf5309_0
jedi=0.18.1=py39hcbf5309_1
joblib=1.1.0=pypi_0
jupyter_client=7.3.0=pyhd8ed1ab_0
jupyter_core=4.9.2=py39hcbf5309_0
kiwisolver=1.4.2=pypi_0
libsodium=1.0.18=h8d14728_1
matplotlib=3.5.2=pypi_0
matplotlib-inline=0.1.3=pyhd8ed1ab_0
nest-asyncio=1.5.5=pyhd8ed1ab_0
numpy=1.22.3=pypi_0
openssl=1.1.1o=h8ffe710_0
packaging=21.3=pyhd8ed1ab_0
pandas=1.4.2=pypi_0
parso=0.8.3=pyhd8ed1ab_0
pickleshare=0.7.5=py_1003
pillow=9.1.0=pypi_0
pip=21.2.4=py39haa95532_0
prompt-toolkit=3.0.29=pyha770c72_0
psutil=5.9.0=py39hb82d6ee_1
pure_eval=0.2.2=pyhd8ed1ab_0
pygments=2.12.0=pyhd8ed1ab_0
pyparsing=3.0.8=pyhd8ed1ab_0
python=3.9.12=h6244533_0
python-dateutil=2.8.2=pyhd8ed1ab_0
python_abi=3.9=2_cp39
pytz=2022.1=pypi_0
pywin32=303=py39hb82d6ee_0
pyzmq=22.3.0=py39he46f08e_2
requests=2.27.1=pypi_0
scikit-learn=1.0.2=pypi_0
scipy=1.8.0=pypi_0
setuptools=61.2.0=py39haa95532_0
six=1.16.0=pyh6c4a22f_0
sklearn=0.0=pypi_0
sqlite=3.38.3=h2bbff1b_0
stack_data=0.2.0=pyhd8ed1ab_0
threadpoolctl=3.1.0=pypi_0
torch=1.11.0=pypi_0
torchvision=0.12.0=pypi_0
tornado=6.1=py39hb82d6ee_3
tqdm=4.64.0=py39haa95532_0
traitlets=5.1.1=pyhd8ed1ab_0
typing-extensions=4.2.0=pypi_0
tzdata=2022a=hda174b7_0
urllib3=1.26.9=pypi_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
wcwidth=0.2.5=pyh9f0ad1d_2
wheel=0.37.1=pyhd3eb1b0_0
wincertstore=0.2=py39haa95532_2
zeromq=4.3.4=h0e60522_1
